+++ smbase/strsobjdict.h
@@ -40,7 +40,7 @@ public:     // types
     string const &key() const { return iter.key(); }
     T * &value() const { return (T * &)iter.value(); }
 
-    int private_getCurrent() const { return iter.private_getCurrent(); }
+    long private_getCurrent() const { return iter.private_getCurrent(); }
   };
   friend class Iter;
 
@@ -59,7 +59,7 @@ public:     // types
     string const &key() const { return iter.key(); }
     T * value() const { return (T *)iter.value(); }
 
-    int private_getCurrent() const { return iter.private_getCurrent(); }
+    long private_getCurrent() const { return iter.private_getCurrent(); }
   };
   friend class IterC;
